[4] Add the first version of the Dashboard
This commit will add the basic structure of the application starting
with the App component which defines the layout of the main parts
of the user interface (Navbar, Aside, Main).

A new component named Main defines the main area of the application
where the various pages will appear thanks to the router. In this main
part, the DashboardView will be used to display the content of the
dashboard. In order to do that, it will use a Card to display various
projects, the ProjectSummaryCard. This ProjectSummaryCard is created
using both the Card and List components.

Unit tests have also been contributed in order to prevent regressions.

Bug: https://github.com/eclipse/sirius-components/issues/4
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>