[6] Display one project
This commit will add a new page displaying the details of one project.
A link has been added from both the dashboard and the list of projects
to a page showing a project.

A project is displayed using the ProjectView component which uses both a
ProjectRepresentationsListCard and a ProjectSemanticResourcesListCard.
On top of the page, the details of the project itself will be visible in
the ProjectHeaderCard component.

Bug: https://github.com/eclipse/sirius-components/issues/6
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>