[7] Connect the front-end with the back-end
This commit will add new stateful components which will be used to hold
the state for the user interface and communicate with the server. The
static data have been removed and the user interface is now computed
from the responses of the server.

The stateful containers, which can not be easily reused have been placed
in a dedicated folder named containers. This way, reusable and non
reusable components can easily be identified.

Bug: https://github.com/eclipse/sirius-components/issues/7
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>