[8] Define properly the states of stateful components
This commit will add finite state machines in order to define all the
states and transitions that the stateful components can use. All the
transitions between the states have been extracted in reducers.

Dispatchers have been introduced in order to ensure that the transitions
and their results are matching the data of the finite state machine of
the component.

Bug: https://github.com/eclipse/sirius-components/issues/8
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>