[11] Define properly the states of stateful components
This commit will modify the various state container components to remove
the rendering logic. This way our stateful components are separated in
three parts. First the behavior in the reducers. Second, the state and
its lifecycle in the state containers and finally the rendering in the
view components.

The state containers and the rendering components are connected by the
Main component now. The rendering components are now way easier to
understand since they can leverage the stateId property in order to show
the various state of the user interface. The whole can thus be reused
much more easily.

Bug: https://github.com/eclipse/sirius-components/issues/11
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>