[9] Handle the error state of the stateful components
This commit will improve the stateful components by handling properly
the error state. Errors can appear because the server can return a
response with an HTTP error status (4xx or 5xx). They can also appear if
an exception occur during the processing of a response from the server.

A new component has been introduced to display error messages to the end
user. This component is now used in all our pages.

Bug: https://github.com/eclipse/sirius-components/issues/9
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>