[28] Update the dashboard will additional data
This commit will add new cards to the dashboard in order to display some
global information regarding the server. It will thus show the number of
projects available along with the number of viewpoints and metamodels.

The layout of the dashboard has been updated with the introduction of a
new kind of card, the InfoCard.

Bug: https://github.com/eclipse/sirius-components/issues/28
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>