[36] Improve the list components' layout
This commit updates the layout of the list components. The "smart" List
Item component has been removed in order to use more explicit components
which are easier to manipulate and understand.

The FooterLink component has also been removed in prevision of the
refactoring of the card components. Some changes in the fonts and sizes
have occurred.

Bug: https://github.com/eclipse/sirius-components/issues/36
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>