[34] Add support for project description
The description of the project is now retrieved from the server. It is
now used in the dashboard instead of using information from the sirius
session which would be expensive to compute. The description is also
visible in the list of projects.

The description can be viewed and edited in the project page thanks to
the project header card.

Bug: https://github.com/eclipse/sirius-components/issues/34
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>