[93] Add support for deployment at a specific path
While building the project, we can now specify the path where the code
will be deployed in order to make the loading of resources and the
router work.

If you want to deploy the code at the path /workflow, you would have to
set the environment variable PUBLIC_URL to /workflow before the build.

Bug: https://github.com/eclipse/sirius-components/issues/93
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>