[130] Display mutation/query errors in diagram UX
Display errors thanks to the Banner component.
Mutation and queries errors in DiagramWebSocketContainer must be
displayed clearly.

In this implementation, we avoid to change the heavy diagram
machine/reducer thanks to a dedicated useState.
This choice can be debated.
The aim of this implementation:handle a basic UX feedback as simple as
possible.
the actual state machine handle too many things/concepts and the purpose
to display "something goes wrong" should not interact with this complex
state machine(complex by it size)

Fixes #130

Signed-off-by: Hugo Marchadour <hugo.marchadour@obeo.fr>