[153] Allow to close opened diagram tabs
A blue cross icon has been added on the diagram tab to be able to close
them.
RAF: Closing the last opened diagram automatically reopens it because of
the representationId kept by the method useParams in EditProjectView.
Also, the history update in that class has been commented as it was
causing to return to the projet page when refreshing a page showing a
diagram.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/153
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>