[241] Remove the dependency to Sirius Web Services API in the diagram part
Sirius Web Spring Collaborative Diagrams does not need to depend on the Sirius Web
Services API anymore thanks to the introduction of two new projects:

- sirius-web-diagrams-services-api
- sirius-web-diagrams-services

These two new projects are used to allow someone else to use a different persistence
strategy for the diagrams. It will need to be refined but this solution allows us to
isolate our dependency to the IRepresentationRepository.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/241
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>