[224] Add Project.currentEditingContext reference
Add the EditingContext entity and repository, along with the
corresponding table in the DB.

Projects get a fresh "currentEditingContext" when they are created.
The editingContext is removed from the database when the project is
deleted.

For now the context is not used for anything, it will be exploited by
later commits.

Bug: eclipse-sirius/sirius-components#224
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>