[224] Move documents into editing context
Instead of being "owned" directly by a project, documents are now
inside an editing context. For basic use cases this will be a
project's current editing context, with no visible difference in
behavior except a new indirection, but this allows documents to also
be owned by modelers (for documents which represent domain definitions
and representation configurations), or (later on) in
non-current (historical) poroject editing contexts.

Bug: eclipse-sirius/sirius-components#224
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>