[cleanup] Make the navbar used explicit
Views are not handling the lifecycle of their navbar and it creates numerous issues
since it complexify the lifecycle of the navbars. On top of that, it degrades the
user experience because everyone starts the creation of a new view by using the
View component without taking into account whether it's relevant or not.

Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>