[357] Prepare the removal of SubscriptionDescription
This first commit will make sure that the disposal of the representation event
processors do not rely on SubscriptionDescription anymore. Thanks to this commit
it will be possible to remove the SubscriptionDescription during the retrieval of
the payload flux and then to update the lifecycle of the APIs used to disposed
the representations event processors.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/357
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>