[fix] Force the use of guava 30
In its plug-in dependencies, ELK requires guava [15,19), but, since a
recent version ELK uses guava Streams which have been introduced in
guava 21.
To prevent ClassNotFoundException to occur because of guava Streams we
force the use of guava 30.

We will be able to revert this commit when ELK 0.8.0 will be released

Signed-off-by: Guillaume Coutable <guillaume.coutable@obeo.fr>