[402] Remove the dependency to GraphQL from the Collaborative Diagram project
We need to prepare the migration of the business specific code from Sirius Components
to Sirius Web. We don't need a fragmented GraphQL API anymore, we can regroup the various
parts of our API in one project. It will be easier to maintain.

This commit brings back the diagram parts of the GraphQL API in the relevant projects

Bug: https://github.com/eclipse-sirius/sirius-components/issues/402
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>