[402] Remove the Diagrams Services and Diagrams Services API projects
Since the business code used to manipulate the persistence layer will move to Sirius Web
We can now inverse the dependency between our collaborative layer and our business layer.
As a result, we can now introduce the generic IRepresentationSearchService API and remove
the dependency to sirius-web-services from the collaborative layer.

These improvements give us the ability to remove the project sirius-web-diagrams-services
and sirius-web-diagrams-services-api which will make the code easier to understand.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/402
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>