[402] Remove most of the dependencies to sirius-web-services and its API
The following projects do not depend on sirius-web-services or sirius-web-services-api
anymore:

- sirius-web-collaborative-api
- sirius-web-collaborative-diagrams-api
- sirius-web-collaborative-forms-api
- sirius-web-collaborative-trees-api
- sirius-web-compatibility
- sirius-web-diagrams-layout
- sirius-web-spring-collaborative
- sirius-web-spring-collaborative-diagrams
- sirius-web-spring-collaborative-forms
- sirius-web-spring-collaborative-trees

As a result, those projects will not depend on our business concepts anymore which will
simplify their integration in other projects. The biggest project still depending on those
projects is sirius-web-emf. This situation will need to change in the future.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/402
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>