[569] Move document related code out of sirius-web-emf
The concept of document, not to be confused with the EMF resource that can be used
to store the data of said document is tightly related to business concerns. Various
applications may used different serialization strategies to persist the editing context
and they may not use this POJO with this specific lifecycle.

As such, this code can be moved to sirius-web-services where it is more relevant.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/569
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>