[569] Remove the last dependencies to sirius-web-services from sirius-web-emf
ObjectService does not need to know about representation or the RepresentationDescriptor. This
concept is coming from our business layer, the EMF part does not need to take care of it.

I have also removed two useless static variables. We need some values but we don't care about the
precise values so no need to share it using global variables.

Finally, the support for custom images is now mostly contained in sirius-web-emf which makes the
odd integration in the lifecycle of the application more visible.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/569
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>