[559] Add Selection Dialog on Composite Processor Tool
When the user selects the Composite Processor Tool in the Flow/Topography Diagram,
a 'Selection Dialog' popup is displayed asking for the selection of an object. The
tool is only executed if the user select an object from the list and clicks on the
'Finish' button, otherwise the tool is not executed.

The 'selection dialog on diagram tools' is an IRepresentation on the backend part.
When a palette tool is executed, the frontend give back to the backend the id of
the element selected in the selection dialog. On the backend side, a new variable
named selectedObject is available in the variable manager if the tool requires a
selection dialog.

The SelectModelElementVariable is supported on both container CreationTool and node
CreationTool. This commit also update the version of Flow Designer used in our tests.
A new tool has been added to Flow-Designer allowing to test SelectModelElementVariable
conversion into SelectRepresentation.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/559
Signed-off-by: Axel RICHARD <axel.richard@obeo.fr>