[562] Move the representation(s) fields from project to editing context
The fields Project#representation and Project#representations are now available
on the type EditingContext instead. It allows us to have features such as the
retrieval of the tools of a diagram or even the onboarding area independant from
the concept of Project.

Most of the references to the concept of project now remain in the mutations.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/562
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>