[559] Create new Selection Representation
Second commit of the 'support for a selection dialog on diagram tools'.

The 'selection dialog on diagram tools' is an IRepresentation on the
backend part.
Create all mandatory plugins for this new kind of IRepresentation.

Also add a test description in 'sirius-web-sample-application', allowing
to test with graphiql that a user can subscribe to this kind of
representation.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/559
Signed-off-by: Axel RICHARD <axel.richard@obeo.fr>