[554] Update properties view initialization
Since properties view is not the only view displayed on the right panel.
When the right panel is switched to the validation view, the properties
view state machine is destroyed. When the user switches back to the
properties view, even if there is a selection the properties view stay
in 'empty' state and wait for another selection.

This commit keeps the initial state selection to undefined in order to
trigger SWITCH_SELECTION at the first render. Doing this has a small
cost, when there is a selection and the user switch from validation view
to properties view, the properties view blink rapidly from the 'empty'
state to the 'ready' state.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/554
Signed-off-by: Guillaume Coutable <guillaume.coutable@obeo.fr>