[DRAFT] Use a Select widget for choosing domain types in view definitions
The new widget is currently on the NodeStyle properties page, for the
simple reason that this one is already customized byu code. Adding the
widget on the actual NodeDescription (and Diagram, and Edge) would
mean re-implementing their whole pages in code.

The basics of the workflow seem to work from simple manual tests, but
there are probably some corner cases to handle.

Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>