[694] Provide a representation to display representations
Projects that depends on Sirius Components should now provide:

- a Spring service that implements IRepresentationsDescriptionProvider
- a Subscription data fetcher to subscribe to the "representationsEvent"

Providing these components will make possible the display of a
representation that display representations for the current selection.

This contribution comes with significant changes to the behavior of the
left and right sides in order to prevent layout issues during the expand
and collapse. It also provides some improvements to the Success object
in order to provide additional information. Those pieces of information
will be used to compute the ChangeDescription. It is a first step to help
users create representations which do not display only semantic elements.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/694
Signed-off-by: Guillaume Coutable <guillaume.coutable@obeo.fr>