[719] Add support for the 'container' and 'element' compatibility variables
'container' is the (default) name for the variable used by Sirius
Desktop to refer to the semantic element on which a node or container
creation tool is invoked. See {Node,Container}CreationCommandBuilder
and ToolFactoryImpl.create{Node,Container}CreationDescription() in
Sirius Desktop.

'element' plays the same role for the generic tools. See
GenericToolCommandBuilder and ToolFactoryImpl.createToolDescription().

Bug: https://github.com/eclipse-sirius/sirius-components/issues/719
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>