[744] Use Reactor to communicate event handler results
Two sinks are now available for the IRepresentationEventProcessor and
their handlers, one to dispatch an IPayload response to the datafetcher,
another one to dispatch a ChangeDescription to the editing context event
processor.

With this change, the IRepresentationEventProcessor should not trigger
their refresh themselves. They will publish a ChangeDescription in the
sink for the EditingContextEventProcessor. Once it has been notified,
it will refresh the representation that has is the source of the change
first, then, he will refresh others representations.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/744
Signed-off-by: Guillaume Coutable <guillaume.coutable@obeo.fr>