[releng] Contribute the first version of the CHANGELOG
The CHANGELOG will be used to communicate with consumers of the Sirius
Components project. It will be maintained by hand by the contributors
since our Git history does not have the level of details required and
it is more oriented toward communication with other developers of the
project. Those developers are quite familiar with the evolution of the
project contrary to external consumers.

Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>