[cleanup] Remove log message from Result
The API is already designed to indicate to the caller if the result
can be interpreted as an int or not. It's up to the client code to
decide what to do if not.

With the log at such a low level, there is no way to avoid "scary"
stack traces in the backend logs in case of a malformed user-supplied
expression for example, which should not be considered even a warning.

Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>