[425] Stick the label to its self-loop edge during the move
We were not able to fix completely the move of elements containing a
self-loop edge. We had to choose between a better UX but not really
reliable (the self-loop edge can go far far away) and a lesser UX but
reliable in any case. We choose the second option, a lesser UX for a
reliable behavior.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/425
Signed-off-by: Guillaume Coutable <guillaume.coutable@obeo.fr>