[fix] Fix the CreateInstance operation when using qualified type names
The previous implementation expected the type name to be the simple
name of an EClass, which should exist in the same EPackage as the
parent/container element. It crashed (unprotected Optional.get() call)
when given a qualified type name (domain::Entity).

Support qualified type names like elsewhere, and use the current
editing context to resolve them into an actual EClass.

Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>