[878] Fix tooltips for tree items
Do not render the raw `kind` as we used to do, as it is now more
confusing than anything for the end-user.

For the kinds which are defined by Sirius Components itself, interpret
them to show the same tooltip as before. Default to an empty tooltip
for the others instead of exposing the raw kind value.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/878
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>