[957] Fix the fit to screen triggered by initialize canvas bounds
When diagram server is created it receives two initialize canvas bounds
actions. The first one is sent at the very start to initialize the
action dispatcher but because the model has not been calculated at this
point the fit to screen is not dispatched here. The second one is
triggered when the diagram server root id has changed. The fit to screen
action is dispatched here.

All other initialize canvas bounds action that come after the fit to
screen action has been dispatched once, will not trigger a fit to
screen. Unless, the user opens a new diagram or switches to another
diagram.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/957
Signed-off-by: Guillaume Coutable <guillaume.coutable@obeo.fr>