[948] Fix edge label positioning
- While computing edge routing points, we consider that an edge from an
element to himself has to be represented by a self loop edge and thus,
it cannot have intersection with its parent bounds.
- Improve the edge label positioning by being more restrictive and it
should be easier to understand that we only compute the label position
of two kinds of edge. Those who are straight between two distinct
elements and self loop edge. But we don't support edge with
bendpoint yet.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/948
Signed-off-by: Guillaume Coutable <guillaume.coutable@obeo.fr>