[779] Add the ability to compute the tools available for a given diagram element
A new API based on the IToolSectionsProvider is now available in order to let
specifiers define which tools are available dynamically when the contextual
palette is requested by the frontend.

An implementation has been provided for the Sirius RCP compatibility layer which
uses the various preconditions which can be defined in an odesign file in order
to compute the tool sections and tools which should be available for a specific
diagram element.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/779
Signed-off-by: Axel RICHARD <axel.richard@obeo.fr>