[956] Improve the layout of border nodes
This commit changes the way the child is added to its parent in convertDiagram.
A child is now added with the add method instead of parent.children=xxx. It
allows to properly set SParentElement.parent property.

Each child adds itself to its parent as soon as created to make sure that its
own children will be added in a graph containing a root element.

The border node enters its parent container with 8px. The ELK layout has been
updated to have a border node offset. The ELK option CoreOptions.PORT_BORDER_OFFSET
is set by default to 8 pixels entering the parent node.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/956
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>