[1071] Support the border node label display
The auto layout uses the CoreOptions.PORT_LABELS_PLACEMENT with the
value PortLabelPlacement.outside(). It means that the label is
positioned on a corner of the border node without entering the parent
node.
The incremental layout uses, for the border nodes label, the same
policy. It is applied when the border node is moved on its container and
when it is newly created.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/1071
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>