[781] Add support for multiline labels
The changes are about the node label center position.
There is in fact no visible change but there are technical changes to
prepare the following commits when we will have a label on many lines:

* the label position given by the server is the horizontal center of the
node instead of the left position of the label
* consequently the LabelView is updated to display the text with
'text-anchor': 'middle' instead of the default 'text-anchor': 'start'

The default form description will now use a text area for default form
representation. It will allow multiple line edition. Shift + return in
a text area only adds a line return but does not send the mutation.

Adapt the ELK to consider the multiple line. The ELK label TextBound is
set according the line returns contained in the label text.
This will allow to have the right width of the node and to avoid the
overlap of the label text with the contained nodes.

Add a isMultiLine attribute to Label so that it can be seen as a text area
by sprotty. Adapt the editing bounds and the edited text font to have always
the same editing area whatever the zoom level.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/781
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>