[1138] Add the ability to minimize the content of a site
- Convert Panel's CSS to Material UI
- Make Panels handle both sides and instanciate the Sites itself
- Enforce a minimum width for panels
- Improve the unit tests and typing of the code
- Prevent text selection aretfacts when resizing panels
- Support expand/collapse of the panels

Panels can be collapsed from the site's icon bar by clicking on the
currently visible view's icon or by resizing to the minimum width.

They can be expanded again by clicking on a view's icon. If the panel
was collapsed by clicking on the icon, the previous width is restored.
If it was collapsed by a "resize to min", the default/initial size of
the panel is used.

- Enforce a minimum width for the main area

Prevent the panels to be enlarged too much so that the main area is
always at least partially visible, and void extreme cases where panels
could overlap with each other causing global overflow.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/1138
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>