[1231] Extract the tree and explorer support in a dedicated package
This commit also improves the support of the existing unit tests in our build.
While the test were previously being compiled successfully, this commit makes
sure that they are also executed with success.

Code coverage results will also be stored for the frontend packages along the
results of the backend packages.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/1231
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>