[perfs] Make RepresentationDescriptionSearchService more efficient
Stop the search as soon as a match is found.

In the common case where the searched representation description is a
"static" one, this completely avoids searching for/loading/converting
the dynamic representations.

Event when searching for a dynamic representation description, stop
loading/converting View models as soon as the target is found.

Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>