[1563] Remove implicit behaviors in View-based diagrams
If a diagram description does not explicitly configure a delete/direct
edit/element creation tool, the corresponding behavior is simply
disabled.

By default, new nodes, edges and tools are explicitly configured to
invoke the default canonical behavior; it is just made explicit in the
model and can now serve as a base or removed to disable the behavior.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/1563
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>