[288] Make sure all handlers see the full version of convertedNodes
In very early versions of `ViewConverter`, the equivalent of the
`convertedNodes` map was stored as instance fields in `ViewConverter`
itself. Because the same instance of `ViewConverter` was used to
convert all DiagramDescriptions inside a given View, each lambda
created during the conversion process of a Diagram got a fixed copy of
that state, which could then be reset before processing the next
diagram to avoid leaking references between unrelated diagrams.

This used to i) be required to avoid leaks and ii) work because we
could fully process (convert) all node descriptions before handling
the actual tools, which is the only part that *referenced* this state.

This no longer works with the introduction of
`NodeDescription.Builder.dropNodeHandler()`, which needs to capture a
reference to the full `convertedNodes` (to expose it as a variable)
_while_ we are converting the NodeDescriptions. The snapshot of
`convertedNodes` exposed in
`ViewDiagramDescriptionConverter.createDropNodeHandler()` is thus
incomplete as not all nodes have yet been converted when this is
invoked.

However, all this snapshotting process is also no longer needed. Since
the introduction of `ViewDiagramDescriptionConverterContext`, there is
no longer any risk of leaking conversion state between different
diagram descriptions. We can directly expose the
`ViewDiagramDescriptionConverterContext.getConvertedNodes()` map while
it is being built. Once a given diagram description is finished being
converted, all the VariableManagers created for it (and only them)
reference the same map, which contains the full conversion mapping.

Because the variables are accessible to user-provided code, we still
only expose an unmodifiable view of the Map to avoid unwanted changes.

Bug: https://github.com/eclipse-sirius/sirius-web/issues/288
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>