[2513] Hide node borders in a list layout
The other solution could have to make node borders overlap by reducing
the size of the node.
But a list laid out node can have compartment or icon label (other kind
of node could be supported later) without knowing it.
With that solution the size of the parent should have been reduced if it
has included compartment, but not with icon label.
We do not want the layout of a node to know the type of its children,
that why we have chosen to hide border instead.

Bug: https://github.com/eclipse-sirius/sirius-web/issues/2513
Signed-off-by: Guillaume Coutable <guillaume.coutable@obeo.fr>