[2515] Fix label height in reactflow diagram
Remove the reset `line-height` property set for some kind of input used
as button.
Since all nodes in react flow are buttons and because the label height
computed during the pre-layout phase was not concerned by the reset
`line-height`, the node height was computed with the height without the
reset `line-height` while in the diagram, it was.

Bug: https://github.com/eclipse-sirius/sirius-web/issues/2515
Signed-off-by: Guillaume Coutable <guillaume.coutable@obeo.fr>