[2484] Keep bottom label only to work
Works only for outside bottom label since outside label are outside only
because of the overflow. A rectangle node is built from two div
elements, the first one, the node, takes all the available space, the
second one, the oustside label, is placed "under" because the first one,
took all the space.

Bug: https://github.com/eclipse-sirius/sirius-web/issues/2484
Signed-off-by: Guillaume Coutable <guillaume.coutable@obeo.fr>