[3087] Fix zoom issues
The ghost card or lane created by react-trello depends on an other component "react-smooth-dnd". This component create the "ghost" with a position "fixed" relative to first container. But when transformation is defined in the hierarchy (needed to implement the zoom feature), the position is relative to the first ancestor without transform property. That why we need to specify the ghost parent to avoid a delta between the ghost card / lane and the mouse.

Bug: https://github.com/eclipse-sirius/sirius-web/issues/3087
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>
Signed-off-by: Florian ROUËNÉ <florian.rouene@obeosoft.com>