[Proto] Try to execute subscription (websocket) in request scope
The purpose here was to try to send requests (thanks to a webclient) to another service from within a subscription with request scope.
From a subscription try to access to RequestContextHolder.getRequestAttributes()

This is very complicated (impossible ?).

This commit contains the work pushed as far as I was capable at the time I made it.
It intends to put calls of ProjectSearchService in a RequestScope, thus, to put calls to it in an execution service that transmit the request scope of the local thread to the editing context thread.
See: DelegatingRequestContextExecutorService and DelegatingRequestContextRunnable
NOTE: Adding the annotation @RequestScope to ProjectSearchService helped to find where to execute things within an execution service.

Thanks to a Handshake interceptor, I try to get the request attribute from the handshake request to reuse them later when the backed receives a websocket message.
The purpose is to try to put the message handling in a request scope.
But, at the time we reuse the request attributes at the message handling, the handshake request is not available anymore, it has been finished.
So that did not work :'(

NOTE: This commit also contains some frontend change to deactivate some subscription that are made outside the editing context to ease the comprehension of the issue

Bug: https://github.com/eclipse-sirius/sirius-web/issues/3677
Signed-off-by: Guillaume Coutable <guillaume.coutable@obeo.fr>