[3958] Prevent Stackoverflow when rendering a twice an expanded TreeItem
This commit provide a partial fix for the given bug. It only prevents
rendering twice in the same tree path and expanded element because it
would cause an infinite loop.

A fix still need to be implemented in the "Expand All" action to prevent
stackoverfow when rendering a infinite loop with non unique ID.

Bug: https://github.com/eclipse-sirius/sirius-web/issues/3958
Signed-off-by: Arthur Daussy <arthur.daussy@obeo.fr>
Signed-off-by: Florian ROUËNÉ <florian.rouene@obeosoft.com>