[3985] Add first support for dark theme
The activated theme is based on the browser preferences for firefox or, user OS preferences for Chrome.
It is based on the `prefers-color-scheme` media query.

The dark theme is not supported everywhere yet.
Styles in Representation descriptions does not leverage on the _color scheme_ yet, icons coming from the backend either.

Bug: https://github.com/eclipse-sirius/sirius-web/issues/3985
Signed-off-by: Guillaume Coutable <guillaume.coutable@obeo.fr>