[4404] Make navigation bar home button fully customizable
* Add 'navigationBarHomeExtensionPoint' which relies on
'navigationBarIconExtensionPoint'. This extension point allows
contributing an arbitrary React component for the top left part of the
navigation bar. It defaults to the same behavior a previously: the
Sirius logo is used as a round button that links to the homepage of the
application.
* Fix corresponding shape and ADR to remove the notion that
'navigationBarIconExtensionPoint' would get deleted.
* Document the new extension point in the changelog. Also add
reference to the shape and ADR that were made for this change.
* Add a few missing type annotations.

Bug: https://github.com/eclipse-sirius/sirius-web/issues/4404
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>