[4376] Work around selection not being set at initial diagram rendering
There was an issue where when resolving a URL with a selection and an
open diagram, the selection was not being applied graphically in the
diagram because at the time the selection is applied, the diagram exists
but has no nodes or edges.

This is a workaround to make sure the selection gets applied during the
first render *after* the nodes and edges have been created.

Bug: https://github.com/eclipse-sirius/sirius-web/issues/4376
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>