[4099] Move proxy validation into DocumentSanitizedJsonContentProvider
We need the actual Resource to perform the validation, so do this
inside DocumentSanitizedJsonContentProvider so that we can cleanup
after ourself and remove the resource used to obtain the initial JSON
content from the resource set.

This is a first step to ensure that we do not have unintended
side-effects on the ResourceSet used, so that later on we can perform
all the upload-related operations inside the already existing
EditingContext's ResourceSet and avoid creating a new only (at great
cost).

Bug: https://github.com/eclipse-sirius/sirius-web/issues/4099
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>