[4099] Cleanup temporary resources in DocumentSanitizedJsonContentProvider
This has a non-trivial cost for medium to large models as removing the
resources triggers
oes.components.emf.services.EditingContextCrossReferenceAdapter.clearReferencesTo(EObject),
which can be quite costly. But it is required to avoid unintended
side-effects if we want to be able to perform the whole upload
operation directly inside the already active EditingContext's
ResourceSet.

Bug: https://github.com/eclipse-sirius/sirius-web/issues/4099
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>