[4886] WIP - Make shareable URL include the query text from the Query view
Notes:
* There's some un-react-y code to initialize the ref of the
WorkbenchViewContributions because we need to be inside a BrowserRouter
to be able to instantiate them.
* In QueryView the ref is typed by WorkbenchViewConfigurationSupplier
when really we return a QueryViewConfigurationSupplier, but I could not
fit that into the type system.
* Also un QueryView we have to cast WorkbenchViewConfiguration into a
QueryViewConfiguration. - Maybe with reified generics we could make that
work (?)
* We can only retrieve the configuration of a view when it is
shown, because otherwise the underlying data does not really exist.

Bug: https://github.com/eclipse-sirius/sirius-web/issues/4886
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>